My son's reality check script needs some work...
Some of us have very large servers, and some people utilize CSF for cloud-based services. Sanity could be seen as a negative term for those with disabilities. You might see a setting for a large server, struggle to believe that they have a better server than you, and claim that they have an insane value. That makes you look incompetent to make a product for people with expensive equipment. Expensive people pay for these fixes.