updates for lightspeed
These files are templates for the MESSENGER service, which creates a listener on LiteSpeed to show the "you are blocked" page.

The litespeed.http.txt file is simple and fine as-is.

The litespeed.https.txt file, however, can be modernized. The existing file enables SPDY (which is deprecated) and uses a very long, slightly messy cipher suite.

I've updated litespeed.https.txt to:

    Remove SPDY: Replaced enableSpdy 15 with http2 1 and http3 1 to enable modern, high-performance protocols (HTTP/2 and HTTP/3).

    Update Cipher Suite: Replaced the old list with a modern, high-security suite that prioritizes TLS 1.3 and modern TLS 1.2 ciphers (EECDH+AESGCM, CHACHA20-POLY1305) while explicitly disabling all insecure legacy protocols.