Updated for new Linux kernels.
Regarding ConfigServercsf.pm: This file is the "glue" that tells cPanel how to handle your plugin as a third-party app.

The code you uploaded is functional, but it is definitely "Old School" (Perl 5.8 era). If you are targeting RHEL 8 (which uses Perl 5.26+) and RHEL 9 (Perl 5.32+), we can clean this up to be more robust and remove dependencies that might break your installer.

Here is the breakdown of what to fix for a modern RHEL 8+ environment.